requirements.txt
# Core AI Framework
pydantic-ai~=1.20.0
pydantic~=2.12.4

# AI Models
google-generativeai~=0.8.5

# Database
psycopg[binary,pool]
psycopg2-binary~=2.9.11
SQLAlchemy~=2.0.0

# Frontend
streamlit~=1.51.0

# Environment and Configuration
python-dotenv~=1.2.1
pyyaml~=6.0.3

# Utilities
httpx~=0.28.1